+++ GameTest/BehaviourTree/DecoratorNode.h
@@ -7,7 +7,7 @@ class DecoratorNode :
 {
 public:
     // template value since decorators can assess different types of values
-    DecoratorNode(int id, T value_to_assess, T supposed_value) : 
+    DecoratorNode(unsigned short id, T value_to_assess, T supposed_value) : 
         IBehaviourNode(id), _value_to_assess(value_to_assess), _supposed_value(supposed_value) {}
 
     void SetValueToAssess(const T& value_to_assess) { _value_to_assess = value_to_assess; }
