+++ GameTest/BehaviourTree/SelectorNode.h
@@ -1,10 +1,11 @@
 #pragma once
 #include "IBehaviourNode.h"
+
 class SelectorNode :
     public IBehaviourNode
 {
 public:
-    SelectorNode(int id) : IBehaviourNode(id) {}
+    SelectorNode(unsigned short id) : IBehaviourNode(id) {}
 
     bool CanHaveChildren() override { return true; }
