+++ GameTest/GameObjects/Path.h
@@ -4,6 +4,7 @@ class Path :
     public Door
 {
 public:
-    Path(std::shared_ptr<Renderer> renderer, float pos_x, float pos_y, TagType tag, int id, char* linked_map) : Door(renderer, pos_x, pos_y, tag, DoorStateType::UNLOCKED, id, linked_map) {}
+    Path(std::shared_ptr<Renderer> renderer, float pos_x, float pos_y, TagType tag, unsigned short id, char* linked_map) : 
+        Door(renderer, pos_x, pos_y, tag, DoorStateType::UNLOCKED, id, linked_map) {}
 };
